src/main/java/dnf/itemauction/domain/member/MemoryMemberRepository.java
package dnf.itemauction.domain.member;

import lombok.extern.slf4j.Slf4j;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import java.util.concurrent.ConcurrentHashMap;

@Slf4j
public class MemoryMemberRepository implements MemberRepository{
    private static Map<Long, Member> memberStore = new ConcurrentHashMap<>();
    private static long sequence = 0L;

    @Override
    public Member save(Member member) {
        member.setId(++sequence);
        memberStore.put(member.getId(),member);
        return member;
    }

    @Override
    public Member findById(Long id) {
        return memberStore.get(id);
    }

    @Override
    public Optional<Member> findByLoginId(String loginId) {
        return findAll().stream()
                .filter(member -> member.getLoginId().equals(loginId))
                .findFirst();
    }

    @Override
    public List<Member> findAll() {
        return new ArrayList<>(memberStore.values());
    }

}
